@@ -14,15 +14,15 @@
     "vue": "^3.5.13"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/preset-env": "^7.15.8",
+    "@babel/core": "^7.26.10",
+    "@babel/preset-env": "^7.26.9",
     "@vue/compiler-sfc": "^3.5.13",
-    "babel-loader": "^8.2.3",
+    "babel-loader": "^8.4.1",
     "css-loader": "^6.5.1",
     "html-webpack-plugin": "^5.6.3",
     "style-loader": "^3.3.1",
     "vue-loader": "^17.0.0",
-    "webpack": "^5.97.1",
+    "webpack": "^5.98.0",
     "webpack-cli": "^6.0.1",
     "webpack-dev-server": "^4.7.1"
   }
